+++ jabgui/src/main/java/org/jabref/gui/entryeditor/JumpToFieldViewModel.java
@@ -1,13 +1,16 @@
 package org.jabref.gui.entryeditor;
 
+import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import javafx.beans.property.SimpleStringProperty;
 import javafx.beans.property.StringProperty;
 
 import org.jabref.gui.AbstractViewModel;
+import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.field.Field;
+import org.jabref.model.entry.field.FieldFactory;
+import org.jabref.model.entry.field.InternalField;
 
 public class JumpToFieldViewModel extends AbstractViewModel {
 
@@ -23,17 +26,23 @@ public class JumpToFieldViewModel extends AbstractViewModel {
     }
 
     public List<String> getFieldNames() {
-        if (entryEditor.getCurrentlyEditedEntry() == null) {
+        BibEntry entry = entryEditor.getCurrentlyEditedEntry();
+        if (entry == null) {
             return List.of();
         }
 
-        return entryEditor.getAllPossibleTabs().stream()
-                          .filter(FieldsEditorTab.class::isInstance)
-                          .map(FieldsEditorTab.class::cast)
-                          .flatMap(tab -> tab.getShownFields().stream())
-                          .map(Field::getName)
-                          .distinct()
-                          .sorted()
-                          .collect(Collectors.toList());
+        return suggestedFields(entry).stream()
+                                     .map(Field::getName)
+                                     .distinct()
+                                     .sorted()
+                                     .toList();
+    }
+
+    private List<Field> suggestedFields(BibEntry entry) {
+        List<Field> suggestedFields = new ArrayList<>();
+        suggestedFields.add(InternalField.KEY_FIELD);
+        suggestedFields.addAll(entry.getFields());
+        suggestedFields.addAll(FieldFactory.getAllFieldsWithOutInternal());
+        return suggestedFields;
     }
 }